+++ src/pages/cart-temp/Cart.module.scss
@@ -1,0 +1,59 @@
+.cart {
+  padding: 3rem;
+  padding-bottom: 11rem;
+}
+
+.checkbox {
+  accent-color: $color-pink-main;
+}
+
+.totalCheckboxContainer {
+  display: flex;
+  justify-content: space-between;
+}
+
+.totalCheckbox {
+  display: flex;
+  gap: 0.8rem;
+  margin-bottom: 3rem;
+  color: $color-gray-600;
+}
+
+.faTrash {
+  color: $color-gray-400;
+}
+
+.noProduct {
+  display: flex;
+  justify-content: center;
+  @include font-body5;
+  color: $color-gray-400;
+  background-color: $color-gray-50;
+  border-radius: 1.2rem;
+  width: 100%;
+  height: 16rem;
+  align-items: center;
+}
+
+.bottomNavCart {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+  bottom: 0;
+  left: 0;
+  box-shadow: 0px -4px 20px 0px rgba(0, 0, 0, 0.08);
+  background-color: $color-white;
+  padding: 2rem;
+  position: fixed;
+  width: 36rem;
+  height: 9.4rem;
+}
+
+.howMuchMinus {
+  text-align: center;
+  font-size: 1.4rem;
+}
+
+.pink {
+  color: $color-pink-main;
+}
